im not sure how or why every 'u32' changed to a 'u64' but hopefully this doesnt mess anything up

+++ plasm2_emu/basetypes.h
@@ -7,7 +7,7 @@ plasm2_emu
 */
 
 typedef unsigned long long u64;
-typedef unsigned long u64;
+typedef unsigned long u32;
 typedef unsigned short u16;
 typedef unsigned char byte, _bool;
